Exchange button invisible when loading or error, validate currency exchange

+++ app/src/main/java/ru/pavelcoder/modulbankdemo/model/bank/TransactionRequest.kt
@@ -3,7 +3,7 @@ package ru.pavelcoder.modulbankdemo.model.bank
 data class TransactionRequest(
     val sourceCurrency: Currency,
     val destinationCurrency: Currency,
-    val amount: Long,
+    val amount: Double,
     val calcFrom: TransactionCalcFrom
 ) {
